pyproject.toml
[project]
name = "tennis-prediction"
description = "MLOps with Databricks course."
requires-python = ">=3.11, <3.13"
dependencies = [
    "mlflow==2.17.0",
    "cffi==1.17.1",
    "cloudpickle==3.1.0",
    "matplotlib==3.9.2",
    "numpy==1.26.4",
    "pandas==2.2.3",
    "psutil==6.0.0",
    "pyarrow==14.0.1",
    "scikit-learn==1.5.2",
    "lightgbm==4.5.0",
    "scipy==1.14.1",
    "databricks-feature-engineering==0.6",
    "databricks-feature-lookup==1.2.0",
    "databricks-sdk==0.32.0",
    "pydantic==2.9.2",
    "loguru==0.7.3",
    "tqdm>=4.67.1",
    "xgboost>=3.0.1",
    "python-dotenv>=1.1.0",
]

dynamic = ['version']

[project.optional-dependencies]
dev = ["databricks-connect==15.4.5",
       "ipykernel>=6.29.5, <7",
       "pip>=24.2",
       "pre-commit>=4.1.0, <5",
       "pytest>=8.3.4, <9",
       "pydantic-settings>=2.8.1",
       "marvelous@git+https://github.com/end-to-end-mlops-databricks-3/marvelous@0.1.0"
    ]
test = [
    "pre-commit>=4.1.0, <5",
    "delta-spark==3.3.0",
    "pyspark==3.5.5",
    "pytest>=8.3.5",
    "pytest-cov >=6.1.0",
    "pydantic-settings>=2.8.1",
    "marvelous@git+https://github.com/end-to-end-mlops-databricks-3/marvelous@0.1.0"
]

[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where =["src"]
include = ["tennis-prediction*"]

[tool.setuptools.dynamic]
version = { file = "version.txt" }

[tool.pytest.ini_options]
testpaths = [ "tests" ]
pythonpath = [ "." ]

# [lint.pydocstyle]
# convention = "sphinx"
